refactor: 공통코드 패키지 변경
