+++ src/main/java/nextstep/subway/common/PageController.java
@@ -1,4 +1,4 @@
-package nextstep.subway;
+package nextstep.subway.common;
 
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
