+++ src/main/java/nextstep/subway/member/domain/Member.java
@@ -1,7 +1,8 @@
 package nextstep.subway.member.domain;
 
-import nextstep.subway.BaseEntity;
 import nextstep.subway.auth.application.AuthorizationException;
+import nextstep.subway.common.BaseEntity;
+
 import org.apache.commons.lang3.StringUtils;
 
 import javax.persistence.Entity;
